fix: Missing getServerSnapshot, which is required for server-rendered content
